c2pa_spec

Type Alias AssertionMetadataAssertion

pub type AssertionMetadataAssertion = AssertionMetadataMap;

Assertion that describes additional information about an assertion, including a hashed-uri reference to it

Aliased Type

pub struct AssertionMetadataAssertion {
    pub date_time: Option<String>,
    pub review_ratings: Option<Vec<RatingMap>>,
    pub reference: Option<HashedUriMap>,
    pub data_source: Option<SourceMap>,
    pub localizations: Option<Vec<LocalizationDataEntry>>,
    pub region_of_interest: Option<RegionMap>,
}

Fields
date_time: Option<String>

The RFC 3339 date-time string when the assertion was created/generated

review_ratings: Option<Vec<RatingMap>>

Ratings given to the assertion (may be empty)

reference: Option<HashedUriMap>

hashed_uri reference to another assertion that this review is about

data_source: Option<SourceMap>

A description of the source of the assertion data, selected from a predefined list

localizations: Option<Vec<LocalizationDataEntry>>

localizations for strings in the assertion

region_of_interest: Option<RegionMap>

describes a region of the asset where this assertion is relevant